# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Fund Management System",
    "description" : "The money surface of a broking account.\n\nTwo conventions hold across every endpoint:\n\n**Money is an integer count of paise, never a decimal.** Rule R5 of\nthe ratified event taxonomy and HLD §9.1c. A float in a published\nschema propagates into every generated client, and those are not\nours to fix.\n\n**Errors are branched on by `code`, never by `message`.** The code\nis stable and safe to send to a client; the message is a\ndeveloper-facing explanation and is not user copy.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "platformAuth" : [ ]
  } ],
...
  "paths" : {
    "/api/v1/funds/payout" : {
      "get" : {
        "tags" : [ "Withdrawals" ],
        "summary" : "The account's open request, if any",
        "description" : "Returns 204 when there is none, which is an ordinary state rather than an error.",
        "operationId" : "openRequest",
        "responses" : {
          "200" : {
            "description" : "The open request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PayoutRequestResponse"
                }
              }
            }
          },
          "204" : {
            "description" : "No open request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PayoutRequestResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Withdrawals" ],
        "summary" : "Create the account's single open withdrawal request",
        "description" : "Rule W3: the request reserves nothing. It is settled at end of day against\nwhatever is available then, which is why the response carries a shrink warning\nkey that must be shown before the trader commits.\n\nRule W4 allows one open request per account, and it is enforced by a partial\nunique index rather than by a check here — a read-then-write would be a race\ndressed as validation.",
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PayoutRequestCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Request accepted",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PayoutRequestResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Malformed request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "`request_already_open` (Rule W4), `withdrawable_unavailable` (the derivation and RMS disagree), or `figures_stale`",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "`amount_exceeds_withdrawable` (carries the figure) or `destination_not_verified`",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "`upstream_unavailable` or `calendar_unavailable`",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/funds/payout/{requestId}" : {
      "delete" : {
        "tags" : [ "Withdrawals" ],
        "summary" : "Cancel a request that has not yet been instructed",
        "description" : "REQ-305. Permitted while ACCEPTED or QUEUED_FOR_RUN — REQ-619 keeps it\navailable after a rail outage, since a trader whose payout was deferred has\nmore reason to want it stopped, not less.\n\nA request belonging to another trader answers 409 `not_cancellable` with\nreason NOT_FOUND rather than 403: confirming existence would itself leak.",
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cancelled",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PayoutRequestResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "`invalid_request` — the id in the path is not a number. Declared because it is reachable: the path variable is a `long`, so an unparseable segment is refused at binding and never reaches the handler",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "`not_cancellable`, with reason ALREADY_INSTRUCTED, ALREADY_TERMINAL or NOT_FOUND",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "MoneyDto" : {
        "type" : "object",
        "description" : "A monetary amount as an integer number of paise. Never a decimal or a float.",
        "properties" : {
          "paise" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Amount in paise. 12345 is ₹123.45.",
            "example" : 12345
          },
          "currency" : {
            "type" : "string",
            "description" : "ISO 4217 currency code.",
            "example" : "INR"
          }
        }
      },
      "PayoutRequestCommand" : {
        "type" : "object",
        "description" : "Create the account's single open withdrawal request.",
        "properties" : {
          "amount" : {
            "$ref" : "#/components/schemas/MoneyDto",
            "description" : "How much to withdraw, in paise."
          },
          "destinationRef" : {
            "type" : "string",
            "description" : "Profile's reference for a verified bank account of this trader.",
            "example" : "acc-4471",
            "minLength" : 1
          }
        },
        "required" : [ "amount", "destinationRef" ]
      },
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The uniform error body. Clients branch on `code`, never on `message`.",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code.",
            "example" : "amount_exceeds_withdrawable"
          },
          "message" : {
            "type" : "string",
            "description" : "Developer-facing explanation. Not user copy."
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Values needed to explain the refusal without re-fetching."
          }
        }
      },
      "PayoutRequestResponse" : {
        "type" : "object",
        "description" : "An accepted withdrawal request. Rule W3: this reserves nothing.",
        "properties" : {
          "requestId" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 4242
          },
          "fmsReference" : {
            "type" : "string",
            "example" : "FMS-2026-0821-4242"
          },
          "arrivalDateQuoted" : {
            "type" : "string",
            "format" : "date",
            "description" : "REQ-303: when the money should arrive."
          },
          "shrinkWarningKey" : {
            "type" : "string",
            "description" : "Rule W3a copy key, shown before commitment.",
            "example" : "WITHDRAWAL_MAY_SHRINK_AT_SETTLEMENT"
          },
          "state" : {
            "type" : "string",
            "example" : "ACCEPTED"
          },
          "withdrawableAtRequest" : {
            "$ref" : "#/components/schemas/MoneyDto",
            "description" : "The withdrawable figure at the moment of request (Rule W11)."
          },
          "destinationMasked" : {
            "type" : "string",
            "description" : "Masked destination. Never the full account number (Profile PR-31).",
            "example" : "••••4471"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "platformAuth" : {
        "type" : "http",
        "description" : "HTTP Basic, which is what the service enforces today.\n\n**Provisional.** Security review MEDIUM-2 is open on this: Basic carries credentials on every request and has no expiry, revocation, rotation or scope. It is here because a filter chain that refuses beats a correct scheme that does not exist, and it is declared here because a specification naming a scheme the service does not accept produces clients that 401 on every call.\n\nThe deployment is expected to place a gateway in front that validates a platform-issued token. When that issuer is known this scheme becomes `bearer`; the requirement name does not change.",
        "scheme" : "basic"
      }
    }
  }
}